55: extract the symbols related to alerts into the newly-created `frontend/src/features/alerts/` directory
