+++ frontend/src/App.tsx
@@ -11,14 +11,14 @@ import {
 import { useSelector, useDispatch } from "react-redux";
 import { v4 as uuidv4 } from "uuid";
 
+import { Alerts } from "./features/alerts/Alerts";
+import {
+  IActionAlertsCreate,
+  alertsCreate,
+} from "./features/alerts/alertsSlice";
+
 import {
   IState,
-  selectAlertsIds,
-  selectAlertsEntities,
-  IActionAlertsCreate,
-  IActionAlertsRemove,
-  alertsCreate,
-  alertsRemove,
   ActionCreateUser,
   createUser,
   ActionIssueJWSToken,
@@ -168,42 +168,6 @@ export const NavigationBar = () => {
       {"<NavigationBar>"}
       {alwaysVisibleLinks}
       {!hasValidToken ? guestUserLinks : loggedInUserLinks}
-    </React.Fragment>
-  );
-};
-
-export const Alerts = () => {
-  console.log(`${new Date().toISOString()} - React is rendering <Alerts>`);
-
-  const alertsIds = useSelector(selectAlertsIds);
-  const alertsEntities = useSelector(selectAlertsEntities);
-
-  const dispatch: Dispatch<IActionAlertsRemove> = useDispatch();
-
-  const handleClick = (
-    alertId: string,
-    e: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    dispatch(alertsRemove(alertId));
-  };
-
-  const alertsDivs = alertsIds.map((aId: string) => (
-    <div key={aId} style={{ color: "red" }}>
-      <button
-        onClick={(e: React.MouseEvent<HTMLButtonElement>) =>
-          handleClick(aId, e)
-        }
-      >
-        Clear alert
-      </button>
-      {alertsEntities[aId].message}
-    </div>
-  ));
-
-  return (
-    <React.Fragment>
-      {"<Alerts>"}
-      {alertsDivs}
     </React.Fragment>
   );
 };
