+++ frontend/src/constants.ts
@@ -1,0 +1,6 @@
+import { IStateAlerts } from "./types";
+
+export const INITIAL_STATE_ALERTS: IStateAlerts = {
+  ids: [],
+  entities: {},
+};
