+++ frontend/src/types.ts
@@ -1,0 +1,9 @@
+export interface IAlert {
+  id: string;
+  message: string;
+}
+
+export interface IStateAlerts {
+  ids: string[];
+  entities: { [id: string]: IAlert };
+}
